requirements.txt
streamlit
torch
transformers
sacremoses
flair
st-annotated-text